introduced folder structure
