+++ Javascript/counting_clouds.js
