+++ Javascript/counting_valleys.js
