+++ Python/counting-mountains.py
