+++ Python/jumping-on-clouds.py
