--- read_phone_numbers.py
@@ -1,14 +1,0 @@
-'''
--no of test cases
--input phone number according to
-dividing format 
--output
-case #1: print out the digits 
-'''
-
-
-#number of test cases
-test_cases = int(input("How many numbers do you want to test?: "))
-accept_numbers = [input() for i in range (test_cases - 1)]
-different_numbers = input().split( )
-
